ajout d'un fichier message dépendant
