+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/SoGameHD/bot_meteo#readme",
   "dependencies": {
+    "axios": "^0.22.0",
     "discord.js": "^13.1.0",
     "dotenv": "^10.0.0",
     "node": "^16.10.0"
