+++ src/index.js
@@ -1,21 +1,9 @@
 const client = require("./client/client.js");
+const message = require("./utils/message.js");
 require("./services/weather.js");
-
-const prefixCmd = '!';
 
 client.on ("ready", () => {
     console.log("🌤 Weather Bot Online ☁");
 });
 
-client.on("message", msg => {
-
-    if(!msg.content.startsWith(prefixCmd) || msg.author.bot) return
-
-    const args = msg.content.slice(prefixCmd.length).trim().split(/ +/);
-    const command = args.shift().toLowerCase();
-
-    if (command === "daylimeteo") {
-        msg.reply("Voici la météo du jour !");
-    }
-
-});
+client.on("message", message);
