+++ src/utils/date.js
@@ -1,2 +1,18 @@
 const days = ['dimanche', 'lundi', 'mardi', 'mercredi', 'jeudi', 'vendredi', 'samedi'];
 const months = ['Jan', 'Fev', 'Mar', 'Avr', 'Mai', 'Jui', 'Jui', 'Aou', 'Sep', 'Oct', 'Nov', 'Dec'];
+
+const getHours = () => {
+    const date = new Date();
+    const hour = date.getHours() < 10 ? '0' + date.getHours() : date.getHours();
+    const min = date.getMinutes() < 10 ? '0' + date.getMinutes() : date.getMinutes();
+    return hour + ':' + min;
+};
+
+const dateToDayHours = (date) => {
+    const day = days[date.getDay()];
+    const hour = date.getHours() < 10 ? '0' + date.getHours() : date.getHours();
+    const min = date.getMinutes() < 10 ? '0' + date.getMinutes() : date.getMinutes();
+    return day + ' ' + hour + ':' + min;
+};
+
+module.exports = getHours, dateToDayHours;
